chore: move files from lib dir to shared
