+++ src/shared/NpmName.ts
